+++ changelog.md
@@ -1,6 +1,6 @@
 - Fixed a major bug where staggered fire logic could prevent weapons from firing under specific conditions.
 - Custom AI: Ship will not vent if it has an opportunity to finish a heavily damaged target.
-- 
+- Custom AI: Improved ship movement algorithm. Now the AI will not approach to larger targets closer than necessary.
 
 Version 1.11.2