+++ changelog.md
@@ -1,3 +1,5 @@
+Version 1.11.3
+
 - Fixed a major bug where staggered fire logic could prevent weapons from firing under specific conditions.
 - Custom AI: Ship will not vent if it has an opportunity to finish a heavily damaged target.
 - Custom AI: Improved ship movement algorithm. Now the AI will not approach larger targets closer than necessary.